move tests, relative phase sweeps and fixes
